feat flie save and load robot
